nfc-ocs-client/requirements.txt
nfcpy
python-osc